Do not delete every single message

+++ app/src/Discord/Interaction/CleaningInteraction.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Afup\Hermes\Discord\Interaction;
 
+use Afup\Hermes\Repository\Event\FindEventByChannel;
 use Discord\Discord;
 use Discord\Parts\Channel\Message;
 use Discord\WebSockets\Event;
@@ -12,6 +13,7 @@ final readonly class CleaningInteraction implements InteractionInterface
 {
     public function __construct(
         private array $adminUserIds,
+        private FindEventByChannel $findEventByChannel,
     ) {
     }
 
@@ -24,6 +26,11 @@ final readonly class CleaningInteraction implements InteractionInterface
 
             if (in_array($message->author->id, $this->adminUserIds)) {
                 return; // user is admin, let him post
+            }
+
+            $event = ($this->findEventByChannel)((int) $message->channel_id);
+            if (null === $event) {
+                return; // no event in channel, no cleaning is required
             }
 
             $message->delete();
